Remove dependency on customizer-preview
For some reason, when customizer-preview is specified as a dependency, the file is loaded twice and a Javascript error is thrown. Removing the dependency appears to resolve the issue and the functionality of the customizer remains ok.

This resolves the issue where you could not close out of the customizer once it was opened.