Wrap new spine_is_sub() with spine_is_subpage()
In 01c98f8dc0, spine_is_subpage was renamed to spine_is_sub, breaking
themes that relied on this function with a fatal error.

This commit adds a wrapper function just in case the change was
intended. We can remove either or in the future when themes are
prepped for it.