Introduce support for an H1 Header template in the builder
This commit provides both front and back end templates for handling
the addition of an H1 section to a page when using the builder.

We've decided to use <section><h1>Text</h1></section>, as this
seems to make the most sense semantically on a larger scale. We
tossed around the idea of using a <header> element, but we can't
really make that decision for whoever is building the page.