Enable the banner / slideshow mechanism included with Make
Make includes a slider by default, using jQuery Cycle2 to manage
the slides on the front end. We can reuse a lot of this code with
some minor changes to the display on the front end.

This first commit sets things up pretty well. We can now edit the
front end template to display something a bit more custom