Remove parent page link from main header
Very often, parent pages are merely place holders for site structure
and have no content, so linking to them results in a blank page. We
need to remove the link for the time being since one of the main
features of the Spine navigation is eliminating the need for
overview/section pages.