Add custom Builder section styles to shrink icons a bit
This preps us for adding a couple other areas in the near future.

At some point I want to do a better job of managing the upstream
CSS with our custom CSS and will likely introduce a grunt process
to help take care of this. For now, we'll append to upstream and
just try to keep track