Use correct `make_` filters rather than old `ttfmake_`
At some point, Make changed all actions and filters to be prefixed
with `make_` instead of `ttfmake_`. There is a back compat converter
in the upstream theme that I'm guessing I missed in the sync to
our build. Rather than trying to maintain backcompat, this adjusts
the current `ttfmake_` filters to be `make_` filters and should
resolve #145