Column overlay in single admin template should have a unique ID
The ID used when outputting the column overlay data in the single
section template matched the overlay date used for the section
itself. This caused only the entry for Column configuration to
appear whether configuration was selected for column or section.

Adding a -1 to the ID mimics what we do in our columns layout and
fixes the issue.

Fixes #158