Introduce BU Navigation filters from the CCB theme
We filter page URLs so that any page saved as a section label
will have a URL attribute of #. This allows for section labels
to be added to a navigational structure without "Overview" being
created by the Spine.

We filter title attributes on all links in the BU Navigation
structure so that we avoid accidental renamings of the "Overview"
text.