Apply the section background to a wrapper rather than the section
If a background image is assigned to a section, we'll assume intent
to create a wrapper and wrap the section in a `<div>` with the
class of `section-wrapper-has-background`. The background images
will continue to be assigned as data attributes, which means the
feature won't be entirely obvious at first but only accessible
via Javascript. I think.