Wrap main body like main header and main footer
We wrap main header and main footer making it easy to target elements
that may be used elsewhere in main like section, a, etc. Adding a
parallel wrapper to the body enables us to target the content area of a
page without fighting with classes applied to the header and footer.