Add filters and actions to support outside section alterations
Introduces one action, `spine_output_builder_section`, to make it
possible for a child theme or plugin to add additional config
options to a section.

Adds corresponding filters to help handle the saving of that data:
* `spine_builder_save_columns`
* `spine_builder_save_banner`
* `spine_builder_save_header`

This allows for data to be captured and possibly saved. It does
not yet allow for anything to be done with the data in our default
templates.