Add support for section IDs during a refactor of columns template
This got slightly out of hand, but appears to clarify some of the
junk that was mushed together in this template. Now, if section
IDs are passed as part of the section data, they will be output
as part of the wrapper (if it exists) or the section itself.

Currently only applicable to the columns layout.