Add a check for empty background and featured images before output
It was possible (I think) for an non-existing image to be assigned
as a featured image to a post somehow. This may be due to import
from another WordPress installation. In these cases, an empty
string is returned when we retrieve the featured image source. We
can get around this by checking for a value before outputting any
HTML for the image.