Provide title options for front page and page for posts
When pages are selected to display as the front page and as the
page for the main posts archive, we should provide a way for those
page titles to display.

This shows options in the customizer for each. These options only
display when static pages are selected. The titles are off by default
to match current behavior and will display when the boxes are checked.