[PR] Change the active menu item class from 'current' to 'active' (#325)
* Change 'current' menu item class to 'active'

* Update documentation

* Use active rather than current and dogeared for BU Navigation

Because of how the Spine turns active into current and dogeared,
we can rip out a bunch of the extra logic managing the BU Nav output

* Use `array_search` to actually remove extra classes

The previous `unset` was not working. :/

* Strip anchor classes added by BU Navigation

BU Navigation adds a series of `level_#` classes and `active`
classes by default. Because we use `active` elsewhere, and very
specifically on the `li` level, we should remove that here.

* Use `array_diff` to remove extra classes

* Use `current-page-parent` in check for current page parent
The `current-page-parent` class is applied any time a menu item is the parent of the current page, but the `current_page_parent` class seems to be applied only when the current page is also in the menu. Checking against the `current-page-parent` class therefore prevents unexpected behavior in cases where a child page is at the same level in the menu as its parent page.

* Rename and refactor variables for clarity, add rules for events

* Update condition with new variable name