Introduce breadcrumb display on top and bottom of pages
If the Breadcrumb NavXT plugin is enabled, an option will show
in the Customizer to allow a site administrator to select whether
breadcrumbs should appear at the top, bottom, or both top and
bottom of a page.

To support this, we finally have a default parts/footers.php file.

Some themes will have already provided this, but probably don't
need to be updated to include the breadcrumb functionality.